基本的な画面の構成をしてみました。パイソンファイルも作ろうとしましたがうまくできませんでした。
