+++ routes/__init__.py
@@ -1,0 +1,2 @@
+# Blueprintをリストとしてまとめる
+blueprints = []
