+++ static/home.css
@@ -1,0 +1,40 @@
+.home_button{
+    position: absolute;
+    height: 100%;
+    background-color: darkgrey;
+}
+
+button{
+    display: block;
+    margin: auto;
+    width: 100px;
+    height: 50px;
+}
+
+.absence{
+    position: absolute;
+    top: 0%;
+    right: 10%;
+    width: 50%;
+}
+
+.absence_list{
+    width: 500px;
+    height: 200px;
+    overflow: scroll;
+    border: black 2px solid;
+}
+
+.todo{
+    position: absolute;
+    top: 50%;
+    right: 10%;
+    width: 50%;
+}
+
+.todo_list{
+    width: 500px;
+    height: 200px;
+    overflow: scroll;
+    border: black 2px solid;
+}
