+++ templates/home.html
@@ -1,0 +1,50 @@
+<!DOCTYPE html>
+<html lang="ja">
+  <head>
+    <meta charset="UTF-8" />
+    <link rel="stylesheet" href="../static/home.css" />
+    <title>ホーム</title>
+  </head>
+  <body>
+    <div class="home_button">
+      <h1>ホーム画面</h1>
+      <button type="button" onclick="loction.href='リンク先のURL'">
+        欠席登録
+      </button>
+      <br />
+      <button type="button" onclick="loction.href='リンク先のURL'">ToDo</button>
+      <br />
+      <button type="button" onclick="loction.href='リンク先のURL'">
+        講義追加
+      </button>
+    </div>
+    <div class="absence">
+      <h2>欠席情報</h2>
+      <div class="absence_list">
+        <!-- ここのリストはテスト用 -->
+        <li>講義名：欠席４回</li>
+        <li>講義名：欠席４回</li>
+        <li>講義名：欠席４回</li>
+        <li>講義名：欠席３回</li>
+        <li>講義名：欠席３回</li>
+        <li>講義名：欠席３回</li>
+        <li>講義名：欠席２回</li>
+        <li>講義名：欠席２回</li>
+        <li>講義名：欠席１回</li>
+        <li>講義名：欠席０回</li>
+        <li>講義名：欠席０回</li>
+      </div>
+    </div>
+    <br />
+    <div class="todo">
+      <h2>ToDoリスト</h2>
+      <div class="todo_list">
+        <!-- ここのリストはテスト用 -->
+        <li>確率統計の課題</li>
+        <li>オブジェクト指向の課題</li>
+        <li>バイト</li>
+        <li>英語</li>
+      </div>
+    </div>
+  </body>
+</html>
