+++ templates/id.html
@@ -1,0 +1,27 @@
+<!DOCTYPE html>
+<html lang="ja">
+  <head>
+    <meta charset="UTF-8" />
+    <link rel="stylesheet" href="../static/id.css" />
+    <title>ホーム</title>
+  </head>
+  <body>
+    <!-- <button type="button" class="id_new" onclick="loction.href='./id_new.html'">
+      ID登録
+    </button> -->
+    <form action="{{ url_for('new') }}">
+      <input type="submit" value="ID登録" />
+    </form>
+    <br />
+    <!-- <button
+      type="button"
+      class="id_login"
+      onclick="loction.href='./id_login.html'"
+    >
+      ログイン
+    </button> -->
+    <form action="{{ url_for('login') }}">
+      <input type="submit" value="ログイン" />
+    </form>
+  </body>
+</html>
