+++ templates/id_new.html
@@ -1,0 +1,23 @@
+<!DOCTYPE html>
+<html lang="ja">
+  <head>
+    <meta charset="UTF-8" />
+    <link rel="stylesheet" href="../static/id_new.css" />
+    <title>欠席登録</title>
+  </head>
+  <body>
+    <h1>新規登録</h1>
+    <form action="home" method="post">
+      <label for="name">ユーザーID:</label>
+      <input type="text" id="name" name="name" required />
+      <br />
+      <label for="pass">パスワード:</label>
+      <input type="text" id="pass" name="pass" required />
+      <br />
+      <input type="submit" value="登録" />
+    </form>
+    <br />
+  </body>
+  <!-- 初期画面はid.htmlのこと -->
+  <a href="/">初期画面に戻る</a>
+</html>
